refactor(android): decompose XmppSessionManager into focused collaborators
1507 lines become a 312-line facade with an unchanged public API over
ActiveSession, ConnectionLoop, XmppEventRouter, ReadStateCoordinator,
OutboundMessenger, ConversationVerbs, SessionCatchup, ResumePersistence,
and a SessionStores aggregate. The 47 existing session-manager tests
pass unmodified; the old fanOut and markConversationDisplayed hotspots
are genuinely split, not moved.
